Fixing double

+++ CIS560Project/CIS560Project/IMovieRepository.cs
@@ -42,6 +42,6 @@ namespace CIS560Project
 
         IReadOnlyList<Movie> GetRecentMovies();
 
-        Dictionary<Movie, int> GetTheaterSales();
+        Dictionary<Movie, double> GetTheaterSales();
     }
 }
